Update verbage on bias example
This might be my American take on it, but the phrase protected groups is confusing to me. In the U.S. at least, there are protected "qualities" like age, sex, national origin, ethnic background, etc., meaning that you cannot discriminate based off of those qualities. I think all groups that are formed along the lines of those qualities are "protected."

Also proposing to add the phrase "that has been historically disenfranchised" to help explain why those biases may exist in a given dataset. (Sidenote: there's definitely a lot of irony in me needing to inject a "why" into a hypothetical statement in this section :) )