Correct quantile definition
This should be the 0.995 quantile, since a quantile always defines how many values are smaller than or equal to it (see https://en.wikipedia.org/wiki/Quantile#Quantiles_of_a_population ). Note that Bau and Zhou also used this correct definition in their paper, but wrote it kind of weird ("select the quantile so that it has 0.5% greater than it"), which I believe lead to the mistake in the book.